# java/api.yaml
swagger: "2.0"
info:
  description: "Planting seeds and watching them grow."
  version: "1.0.0"
  title: "VMware Bifröst Garden"
  contact:
    email: "dshanley@vmware.com"
host: "localhost:3000"
basePath: "/api"
tags:
- name: "seed"
  description: "Seed operations"
  externalDocs:
    description: "Gets grow some stuff!"
    url: "http://vmware.com"
schemes:
- "http"
paths:
  /seed:
    get:
      tags:
      - "seed"
      summary: "Get a list of seeds growing."
      description: ""
      operationId: "getSeeds"
      produces:
      - "application/json"
      parameters: []
      responses:
        default:
          description: "success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Seed"
    post:
      tags:
      - "seed"
      summary: "Plant a new seed"
      description: ""
      operationId: "plantSeed"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Plant a new seed in the garden"
        required: true
        schema:
          $ref: "#/definitions/Seed"
      responses:
        200:
          description: "Newly planted seed represented by the system"
          schema:
            $ref: "#/definitions/Seed"
    delete:
      tags:
      - "seed"
      summary: "Kill a plant"
      description: ""
      operationId: "killPlant"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Kill a plant from the garden"
        required: true
        schema:
          $ref: "#/definitions/Seed"
      responses:
        200:
          description: "success"
        400:
          description: "invalid request"
        404:
          description: "plant / seed not found"

definitions:
  Seed:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      planted:
        type: "integer"
        format: "int64"
        readOnly: true
      type:
        type: "string"
        description: "Type"
        enum:
        - "Tree"
        - "Flower"
        - "Bush"  
    xml:
      name: "Seed"      